issue7: split composite requests into new file composite.go, create baseline tests for composite.go and dml.go
